Intelligently toggle barmoji in predictive bar
When barmoji replaces the predictive bar, it also blocks your ability to "cancel" autocorrections, or make use of autifill. These changes make barmoji hide itself when an autocorrection is about to take place, or when autofill has something to suggest.

This should also resolve #22, since the only reason someone might be annoyed at barmoji appearing with the number pad is if it's blocking an autofill SMS code or something—and now that won't happen.